Named constants via `def` working

+++ src/interpreter.rs
@@ -37,59 +37,82 @@ fn analyze(input: Vec<Form>) -> Result<Analysis, Box<Error>> {
 }
 
 fn compile(input: Analysis) -> Result<Program, Box<Error>> {
-    fn emit_const(p: &mut ProgramBuilder, val: i64) -> Result<(), Box<Error>> {
-        p.load_const(val);
-        Ok(())
+    fn emit_const(p: &mut ProgramBuilder, val: i64) -> Result<bool, Box<Error>> {
+        p.load_i64(val);
+        Ok(true)
     }
 
-    fn emit_var(p: &mut ProgramBuilder, var_id: u32, env: &SymbolTable) -> Result<(), Box<Error>> {
+    fn emit_var(p: &mut ProgramBuilder,
+                var_id: u32,
+                env: &SymbolTable)
+                -> Result<bool, Box<Error>> {
         let var_record = env.lookup_id(var_id).unwrap();
 
-        unimplemented!()
+        match *var_record.kind() {
+            VarKind::Constant => {
+                p.load_named_constant(var_id);
+                Ok(true)
+            }
+            _ => Ok(false),
+        }
     }
 
     fn emit_do(p: &mut ProgramBuilder,
                env: &SymbolTable,
                args: &[AstNode])
-               -> Result<(), Box<Error>> {
+               -> Result<bool, Box<Error>> {
         let (last, others) = args.split_last().expect("emit_do");
 
         for arg in others {
-            try!(emit(arg, env, p));
-            p.drop();
+            // if something was left on the stack, clean it up
+            if try!(emit(arg, env, p)) {
+                p.drop();
+            }
         }
         emit(last, env, p)
     }
 
     fn emit_def(p: &mut ProgramBuilder,
                 env: &SymbolTable,
-                var_record: &Record,
                 args: &[AstNode])
-                -> Result<(), Box<Error>> {
-        //
-        unimplemented!()
+                -> Result<bool, Box<Error>> {
+        assert_eq!(args.len(), 2);
+
+        match args[0] {
+            AstNode::Var(id) => {
+                match args[1] {
+                    AstNode::Const(val) => {
+                        p.def_const(id, val);
+                    }
+                    _ => unimplemented!(),
+                }
+            }
+            _ => return Err("1st arg of `def` must be a var".into()),
+        }
+
+        Ok(false) // def never puts anything on the stack (TODO: nil?)
     }
 
     fn emit_inv(p: &mut ProgramBuilder,
                 callee: &AstNode,
                 args: &[AstNode],
                 env: &SymbolTable)
-                -> Result<(), Box<Error>> {
+                -> Result<bool, Box<Error>> {
 
 
         match *callee {
             AstNode::Var(id) => {
                 let var_record = env.lookup_id(id).unwrap();
                 match *var_record.kind() {
-                    VarKind::Normal => {
+                    VarKind::FnNormal => {
                         for arg in args {
                             try!(emit(arg, env, p));
                         }
                         emit(callee, env, p)
                     }
-                    VarKind::Special => {
+                    VarKind::FnSpecial => {
                         match var_record.ident() {
-                            "def" => emit_def(p, env, var_record, args),
+                            "def" => emit_def(p, env, args),
                             "do" => emit_do(p, env, args),
                             "+" => {
                                 assert!(args.len() < 256);
@@ -99,11 +122,12 @@ fn compile(input: Analysis) -> Result<Program, Box<Error>> {
                                 }
 
                                 p.add(args.len() as u8);
-                                Ok(())
+                                Ok(true)
                             }
                             _ => unimplemented!(),
                         }
                     }
+                    VarKind::Constant => Err("cannot invoke constant".into()),
                 }
             }
             _ => {
@@ -115,7 +139,7 @@ fn compile(input: Analysis) -> Result<Program, Box<Error>> {
         }
     }
 
-    fn emit(node: &AstNode, env: &SymbolTable, p: &mut ProgramBuilder) -> Result<(), Box<Error>> {
+    fn emit(node: &AstNode, env: &SymbolTable, p: &mut ProgramBuilder) -> Result<bool, Box<Error>> {
         match *node {
             AstNode::Const(val) => emit_const(p, val),
             AstNode::Var(id) => emit_var(p, id, env),
@@ -124,7 +148,8 @@ fn compile(input: Analysis) -> Result<Program, Box<Error>> {
     }
 
     let mut program = new_program();
-    try!(emit(input.ast(), input.env(), &mut program));
+    assert!(try!(emit(input.ast(), input.env(), &mut program)),
+            "nothing will be on the stack");
     program.exit();
 
     let program = program.finish();
@@ -150,5 +175,15 @@ mod test {
     #[test]
     fn nested_add_int_constants() {
         assert_eq!(interpret("(+ (+) (+ 1 (+ 1 1)) (+ 1 2))"), "6");
+    }
+
+    #[test]
+    fn def_const_var() {
+        assert_eq!(interpret("(def x 7) x"), "7");
+    }
+
+    #[test]
+    fn add_const_vars() {
+        assert_eq!(interpret("(def x 3) (def y 4) (+ x 7 y)"), "14");
     }
 }
